fix: improve tests a bit (refactoring + rounding)
